[codeql] Widen 16 int-multiplication-before-cast sites to avoid overflow
CodeQL's cpp/integer-multiplication-cast-to-long flagged 16 new sites
where two narrower operands (unsigned int, or a uint32_t returned by
align()/ceilDiv()) are multiplied and the unsigned-int result is then
implicitly converted to a wider type -- size_t, int64_t, or a vector's
size_type -- for an allocation size, buffer offset, or element count.
On dimensions large enough to overflow 32 bits, the multiply wraps
before the widening happens, so the destination sees a truncated value
instead of the true count.

Each site now casts one operand to the destination width before the
multiply: nntrainer/layers/llm/tie_word_embedding.cpp (per-row weight
offset), nntrainer/tensor/float_tensor.cpp (broadcast stride offset),
nntrainer/tensor/int4_utils.cpp (padded scales buffer size), and the
row/column products test/unittest/unittest_nntrainer_cpu_backend.cpp
builds vectors and calls quantization helpers with.

+++ nntrainer/layers/llm/tie_word_embedding.cpp
@@ -241,8 +241,8 @@ void TieWordEmbedding::incremental_forwarding_embedding(
         throw std::invalid_argument("input word index is greater than in_dim");
       }
 
-      nntrainer::Tensor cur_weight =
-        weight.getSharedDataTensor(out_tensor_dim, out_dim * embed_idx);
+      nntrainer::Tensor cur_weight = weight.getSharedDataTensor(
+        out_tensor_dim, static_cast<size_t>(out_dim) * embed_idx);
       nntrainer::Tensor out_tensor =
         batchsliced_hidden.getSharedDataTensor(out_tensor_dim, out_dim * (i));
